requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.1
distro==1.9.0
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
frozenlist==1.6.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.9.0
jsonfield==3.1.0
multidict==6.4.3
openai==1.76.0
packaging==25.0
propcache==0.3.1
pydantic==2.11.3
pydantic_core==2.33.1
python-dotenv==1.1.0
requests==2.32.3
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
whitenoise==6.9.0
yarl==1.20.0
Django>=5.2
whitenoise>=6.4
django-cors-headers>=3.15
gunicorn>=23.0.0
# ... 나머지
